メモリレイアウトを調整

+++ src/spatial_id/collection/flex_tree/core/node_ops.rs
@@ -331,7 +331,7 @@ where
 
         SharedNode::new(Node::Branch {
             level,
-            leaf_count: new_lower.leaf_count() + new_upper.leaf_count(),
+            leaf_count: (new_lower.leaf_count() + new_upper.leaf_count()) as u32,
             max_zoom: Self::fold_max_zoom(level, &new_lower, &new_upper),
             lower_child: new_lower,
             upper_child: new_upper,
