+++ src/spatial_id/collection/flex_tree/core/overlap.rs
@@ -230,7 +230,7 @@ where
                 }
             }
 
-            *leaf_count = lower_child.leaf_count() + upper_child.leaf_count();
+            *leaf_count = (lower_child.leaf_count() + upper_child.leaf_count()) as u32;
             *max_zoom = Node::<V>::fold_max_zoom(*level, lower_child, upper_child);
         }
